requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
dj-database-url==3.0.1
Django==5.2.3
django-appconf==1.1.0
django-compressor==4.5.1
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.9.0
